chg to float; update comment

+++ marathon.c
@@ -7,10 +7,10 @@
 int main()
 {
   int miles = 26, yards = 385;
-  double kilometres;
+  float kilometres;
 
-  /* use 1760.0 to cast to double; otherwise int/int will return 0 */
+  /* use 1760.0 to cast to float; otherwise int/int will return 0; e.g. 385/1760 */
   kilometres = 1.609 * (miles + yards / 1760.0);
-  printf("\nA marathon is %lf kilometres.\n\n", kilometres);
+  printf("\nA marathon is %f kilometres.\n\n", kilometres);
   return 0;
 }
